restructure files for 2017 to 2018 transition
