+++ scripts/measures/2017/add-registry-submission-method.js
