+++ scripts/measures/2017/get-inverse-measures-from-pdfs.js
