+++ scripts/measures/2017/get-quality-measures-from-pdfs.js
