+++ scripts/measures/2017/get-strata-and-uuids-from-ecqm-zip.js
