+++ scripts/measures/2017/get-stratification.js
