+++ test/scripts/measures/2017/import-qcdr-measures-spec.js
@@ -7,29 +7,30 @@ const path = require('path');
 const _ = require('lodash');
 
 // Test data
-const testMeasures = '../../test/scripts/measures/fixtures/test-measures-data.json';
-const testMeasures2 = '../../test/scripts/measures/fixtures/test-measures-data2.json';
-const testCsv = '../../test/scripts/measures/fixtures/test-qcdr.csv';
-const testCsv2Cols = '../../test/scripts/measures/fixtures/test-qcdr-2cols.csv';
-const qcdrStrataNamesFile = '../../util/measures/qcdr-measures-strata-names.json';
-const outputArg = '../../test/scripts/measures/fixtures/test-measures-data-output.json';
-const output = '../' + outputArg;
+const year = 2017;
+const testMeasures = '../../../test/scripts/measures/' + year + '/fixtures/test-measures-data.json';
+const testMeasures2 = '../../../test/scripts/measures/' + year + '/fixtures/test-measures-data2.json';
+const testCsv = '../../../test/scripts/measures/' + year + '/fixtures/test-qcdr.csv';
+const testCsv2Cols = '../../../test/scripts/measures/' + year + '/fixtures/test-qcdr-2cols.csv';
+const qcdrStrataNamesFile = '../../../util/measures/' + year + '/qcdr-measures-strata-names.json';
+const outputArg = '../../../test/scripts/measures/' + year + '/fixtures/test-measures-data-output.json';
+const outputPath = path.join(__dirname, year.toString(), '../../' + outputArg);
 
 // Expected new measures
-const expectedMeasures = require('./fixtures/expected-measures.json');
+const expectedMeasures = require('../' + year + '/fixtures/expected-measures.json');
 
 // Function which executes script and reads in output file to a JS object.
 const runTest = function(measuresFile, measuresCsv) {
-  const cmd = 'node ./scripts/measures/import-qcdr-measures.js ' +
+  const cmd = 'node ./scripts/measures/' + year + '/import-qcdr-measures.js ' +
     measuresFile + ' ' + measuresCsv + ' ' + qcdrStrataNamesFile + ' ' + outputArg;
   console.log(execSync(cmd, {stdio: 'pipe'}).toString());
 
-  const qpp = fs.readFileSync(path.join(__dirname, output), 'utf8');
+  const qpp = fs.readFileSync(outputPath, 'utf8');
 
   return JSON.parse(qpp);
 };
 
-describe('import measures', function() {
+describe(year + ' import measures', function() {
   it('should create new measures and ignore duplicate measureIds', () => {
     const measures = runTest(testMeasures, testCsv);
     assert.equal(measures.length, 3);
@@ -72,8 +73,8 @@ describe('import measures', function() {
 
 after(function() {
   try {
-    fs.unlinkSync(path.join(__dirname, output));
+    fs.unlinkSync(outputPath);
   } catch (err) {
-    console.log('Warning: ', output, ' should have been deleted but was not.');
+    console.log('Warning: ', outputPath, ' should have been deleted but was not.');
   }
 });
