+++ util/measures/2017/additional-stratifications.json
