+++ util/measures/2017/qcdr-measures-strata-names.json
