+++ util/measures/2017/quality-measures-strata-details.json
