+++ util/measures/2018/cpc+-measure-groups.json
@@ -1,0 +1,5 @@
+{
+  "A": ["CMS122v5", "CMS159v5", "CMS165v5"],
+  "B": ["CMS137v5", "CMS139v5", "CMS149v5", "CMS156v5"],
+  "C": ["CMS124v5", "CMS125v5", "CMS130v5", "CMS131v5", "CMS138v5", "CMS166v6", "CMS50v5"]
+}
